--- a/src/kex_rlwe_newhope/kex_rlwe_newhope.c
+++ b/src/kex_rlwe_newhope/kex_rlwe_newhope.c
@@ -153,6 +153,8 @@ void OQS_KEX_rlwe_newhope_alice_priv_free(UNUSED OQS_KEX *k, void *alice_priv) {
 
 void OQS_KEX_rlwe_newhope_free(OQS_KEX *k) {
 	if (k) {
+		free(k->named_parameters);
+		k->named_parameters = NULL;
 		free(k->method_name);
 		k->method_name = NULL;
 	}