The previous implementation of frodo took advantage of having defined constants. This implementation passes in parameters instead.
By caching some parameters that are used in many for loops, performance is greatly increased.