@mofojed 2016-10-31
The performance benchmarks before and after on my machine. Most significant improvement is in the alice 0 step.
Before:
Operation                           Iterations  Total time (s)  Time(us): mean      pop. stdev    Cycles: mean      pop. stdev
LWE Frodo recommended
alice 0                                    278           1.002        3603.043         355.543        10065444          993046
bob                                        593           1.001        1687.992         197.997         4715298          552922
alice 1                                   5995           1.000         166.817          33.965          465921           94813

After:
Operation                           Iterations  Total time (s)  Time(us): mean      pop. stdev    Cycles: mean      pop. stdev
LWE Frodo recommended
alice 0                                   1182           1.000         846.417          75.907         2364544          211965
bob                                        750           1.001        1334.513         129.636         3727863          362115
alice 1                                   7702           1.000         129.842          18.663          362641           52097