@dstebila 2016-11-01
I tried doing an alternate version where I use preprocessor macros so more stuff could be unrolled.  Could you take a look at #45 and let me know what you think?