@dstebila 2016-11-04
Resolved by #45.